Payroll and PayrollDetails Create and Read

+++ SalCentral.Api/DTOs/PayrollDTO.cs
@@ -6,7 +6,7 @@ namespace SalCentral.Api.DTOs
     public class PayrollDTO
     {
         public Guid? PayrollId { get; set; }
-        public double? TotalAmount { get; set; }
+        public decimal? TotalAmount { get; set; }
         public Guid? GeneratedBy { get; set; }
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
