+++ SalCentral.Api/Models/PayrollDetails.cs
@@ -8,9 +8,9 @@ namespace SalCentral.Api.Models
         public Guid PayrollDetailsId { get; set; }
         public Guid PayrollId { get; set; }
         public Guid UserId { get; set; }
-        public double DeductedAmount { get; set; }
-        public double NetPay { get; set; }
-        public double GrossSalary { get; set; }
+        public decimal DeductedAmount { get; set; }
+        public decimal NetPay { get; set; }
+        public decimal GrossSalary { get; set; }
         public DateTime PayDate { get; set; }
     }
 }
